fix: updates logic to get release information (#437)
* fix: updates logic to get release information

Using the gh CLI requires the repository to already be cloned.
The publish.yml clones the repository based on input, so the
latest release is updated to pull the information in a way that
does not require a local copy of the repo.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: address PR feedback

Add "-r" to remove extra quotes to LATEST_RELEASE
in publish.yml

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

---------

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>