chore: simplify pull request template
The type of change is already expressed by the titles. The checklist
could be interesting for new contributors, but this could better fit
project documentation such as Style Guide and CONTRIBUTING.md. We can
also introduce automated tests to ensure adherence. So this PR makes
the new PRs process simpler for our contributors.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>