ci: ensure release pr is triggered manually
Currently we have a release cadence of about three weeks and in the
meantime it is not necessary to keep a release PR open consuming CI
resources whenever a PR is merged into main. When we are closer to a
release, we can manually trigger this workflow. This will also allow CI
tests to be executed on it.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>